Fix fringe case error and change evolution relationship
There was an error where when a flavor text is being searched for, it was not accounting for Meltan & Melmetal being marked as gen 7 whilst not being in Ultra Sun. This resulted in these two not getting their flavor text and crashing the database creation.

The database relationship from Pokemon to Pokemon that was used for indicating what Pokemon a Pokemon evolved from has been changed to one to many because it was not possible to set two Pokemon to potentially be the evolution of a single Pokemon. For example, Wurmple evolves into both Silcoon and Cascoon.